tg_bot/requirements.txt
googletrans==4.0.0rc1
asyncio==3.4.3
nest-asyncio==1.5.8
pyppeteer==1.0.2
beautifulsoup4==4.11.1
async-timeout==4.0.2
cloudpickle==2.0.0
pickleshare==0.7.5
aiogram==2.15
pandas==1.5.3
tabulate==0.8.9
numpy==1.23.5
scikit-learn==1.0.2
psycopg2-binary==2.9.9
aiohttp==3.8.6